lm_ots: Refactor signing for fast verification
The optimize_message_hash_fast_verify fn stores the message_randomizer in a
slice which is split from the message. This enables the return of the
message_randomizer back to the signer such that the message is directly
changed.

Therefore, the sign_fast_verify fn splits the message into the original
message and the message_randomizer. This only happens if
MAX_HASH_OPTIMIZATIONS is not zero and if the message is long enough to
possibly contain the message_randomizer. This is done to allow the
signing of shorter messages where the signer does not want a
message_randomizer.

Before using the message_randomizer slice it is checked if it is empty,
i.e. contains only 0s, to not overwrite parts of the original message.
This check currently does not raise an error. A "regular" signature is
generated if there is no designated space for the message_randomizer.