fast_verify: Speedup message hash optimize 3/3
The optimize_message_hash_fast_verify function is optimized to generate
a new message_randomizer with an hash algorithm instead of the get_random
function. The get_random function is used to initially seed this PRF. To
generate new message randomizers the hash algorithm is fed with the exisiting
message_randomzer as input.

Further is the evaluation of the generated message_hash with
message_randomizer no longer evaluated with an own implementation, but
the existing checksum calculation is reused for this. To further optimize
the evaluation, the required coefficients are calculated once with the
checksum_cached_init function. The cached parameters can then be used with
the checksum_cached function.

Benchmarks for comparisons were run with `cargo flamegraph`:
$ MAX_HASH_OPTIMIZATIONS=50000000 cargo flamegraph --example lms-demo \
    --features fast_verify -- sign keypair message