Refactor fast_verify eval function
The fast_verify eval function was renamed to fast_verify_eval_init &
fast_verify_eval to reflect that no longer the checksum value is used
to evaluate the effectiveness of an optimization. Instead message and
checksum hash chain lengths are added up.

Thus, the sign_fast_verify function has to be changed as well. Instead
of optimizing to a minimum value for the checksum, a maximum value for
the overall hash chain length is the goal.