hss: Split sign fn into sign, *_mut & *_core fn
The new hss_sign_core method is called by the sign & and sign_mut
methods. It forwards both the message & message_mut option to the
HssSignature::sign function.

The public functions are now extented to the sign & sign_mut method, which
only require the necessary arguments: message, private_key,
private_key_update_function and aux_data.

In the case of the sign_mut method, the message must be a mutable
reference. This allows the LMS library to append the nonce for fast
verifiable signatures.