Add multi-threading for fast_verify feature
The optimization count is split into equal parts for each thread. Each
thread choses a random seed for the message_randomizer. After each
evaluation the message_randomizer is hashed and the digest is used as
new message_randomizer for the next trial. The best message_randomizer
is stored together with the overall hash chain length. After finishing
the desired amount of optimizations the result of the best
message_randomizer and the message_randomizer itself is reported back to
the main program. The main program then iterates over these results and
returns the best from all threads.