hss/sign: Delete code for key regenerate
This can be deleted, as the current architecture of the library cannot
sign multiple times with a loaded key pair.

After each sign operation the key is updated. Prior to the next sign
operation the key needs to be passed to the public functions (hss_sign &
*_mut). In these functions the key is parsed and trees are regenerated
if necessary.

The test reuse_loaded_keypair was added to ensure that an invalid reuse
panics.