Extend tests wrt constrained parameter set
- lms/parameters: Add testing parameter with a height of 2 to reduce
computational effort for testing

- hss/rfc: exhaust_state: Checks whether the state including the seed
is wiped, if the total count of signatures was reached.

- hss/rfc: parse_exhausted_state: Checks whether parsing of an exhausted
state panics.

- hss/mod: update_keypair: Keypair before and after signing is compared.
If the update works properly the keypairs are different.

- hss/mod: exhaust_keypair: Exhausts a keypair to compare whether the
private key is correctly incremented and if it is wiped after the total
count of signatures is reached.

- hss/mod: use_exhausted_keypair: An exhausted keypair is used as an
input for signing. This should raise an error.