Refactor RFC private key increment()
The tree heights were passed as a slice reference. This lead to repeated
code to calculate this slice at several positions. Further, the argument
was not very descriptive or restrictive. To enhance this a HssPrivateKey
needs now to be passed to the increment function. The increment function
itself calculates then with HssPrivateKey variable the tree heights
slice.